% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{checkNested}
\alias{checkNested}
\title{Check if groups are properly nested in hierarchy}
\usage{
checkNested(groups, L)
}
\arguments{
\item{groups}{CharacterMatrix defining the group hierarchy structure}

\item{L}{Integer specifying the number of levels in the hierarchy}
}
\description{
Check if groups are properly nested in hierarchy
}
