% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTL_mapping.R
\name{eQTL_mapping_step}
\alias{eQTL_mapping_step}
\title{eQTL Mapping Step}
\usage{
eQTL_mapping_step(
  SNP_file_name,
  snps_location_file_name,
  expression_file_name,
  gene_location_file_name,
  context,
  out_dir,
  output_file_name_cis = if (identical(context, "shared")) {
     file.path(out_dir,
    "shared.cis_pairs.txt")
 } else {
     file.path(out_dir, paste0(context,
    "_specific.cis_pairs.txt"))
 },
  output_file_name_tra = if (identical(context, "shared")) {
     file.path(out_dir,
    "shared.trans_pairs.txt")
 } else {
     file.path(out_dir, paste0(context,
    "_specific.trans_pairs.txt"))
 },
  method = "MatrixEQTL",
  use_model = modelLINEAR,
  cis_dist = 1e+06,
  pv_threshold_cis = 1,
  pv_threshold_tra = 0,
  error_covariance = numeric()
)
}
\arguments{
\item{SNP_file_name}{Full path to SNP genotype matrix.}

\item{snps_location_file_name}{Full path to SNP location file.}

\item{expression_file_name}{Full path to expression matrix.}

\item{gene_location_file_name}{Full path to gene location file.}

\item{context}{Context name for labeling output.}

\item{out_dir}{Output directory.}

\item{output_file_name_cis}{Path to write cis-eQTL output.}

\item{output_file_name_tra}{Path to write trans-eQTL output.}

\item{method}{Either "MatrixEQTL" or "tensorqtl".}

\item{use_model}{MatrixEQTL model (default: modelLINEAR).}

\item{cis_dist}{Distance threshold for cis-eQTLs.}

\item{pv_threshold_cis}{P-value threshold for cis.}

\item{pv_threshold_tra}{P-value threshold for trans.}

\item{error_covariance}{Covariance matrix (or numeric()).}
}
\value{
Writes cis-eQTLs and trans-eQTLs (optional) to file.
}
\description{
Function to map cis eQTLs - cis window is defined as 1Mb
}
\examples{
data_dir <- file.path(tempdir(), "")
simulate_data(
  data_dir = data_dir, N = 10, n_genes = 5,
  n_snps_per_gene = 10, n_contexts = 3, seed = 42
)
decomposition_step(
  exp_mat_filename = file.path(data_dir, "expression.txt"),
  data_dir = data_dir
)
eQTL_mapping_step(
  SNP_file_name = file.path(data_dir, "SNPs.txt"),
  snps_location_file_name = file.path(data_dir, "snpsloc.txt"),
  expression_file_name = file.path(data_dir, "context1_specific_expression.txt"),
  gene_location_file_name = file.path(data_dir, "geneloc.txt"),
  context = "context1",
  out_dir = data_dir,
  output_file_name_cis = file.path(data_dir, "context1_specific.cis_pairs.txt"),
  output_file_name_tra = file.path(data_dir, "context1_specific.trans_pairs.txt")
)

}
