% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_eGenes_multi_tissue_mod}
\alias{get_eGenes_multi_tissue_mod}
\title{modified TreeQTL function}
\usage{
get_eGenes_multi_tissue_mod(
  m_eqtl_out_dir,
  treeQTL_dir,
  tissue_names,
  level1 = 0.05,
  level2 = 0.05,
  level3 = 0.05,
  exp_suffix
)
}
\arguments{
\item{directory}{where matrixeQTL output is stored, directory for treeQTL output, names of contexts}
}
\value{
outputs eAssociation and eGene files with FDR corrected p-values
}
\description{
Function to perform hierarchical multiple testing correction across 3 different levels
}
