% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeBH.R
\name{get_TreeBH_selections_cpp}
\alias{get_TreeBH_selections_cpp}
\title{C++ TreeBH Implementation using Rcpp}
\usage{
get_TreeBH_selections_cpp(pvals, groups, q, test = "simes")
}
\description{
C++ TreeBH Implementation using Rcpp
}
\keyword{internal}
