% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeBH.R
\name{get_TreeBH_selections_datatable}
\alias{get_TreeBH_selections_datatable}
\title{Optimized TreeBH Implementation using data.table}
\usage{
get_TreeBH_selections_datatable(pvals, groups, q, test = "simes")
}
\description{
Optimized TreeBH Implementation using data.table
}
\keyword{internal}
