% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getGroupPvalues}
\alias{getGroupPvalues}
\title{Hierarchical Group P-value Calculation
The getGroupPvalues function computes aggregated p-values for each group in a hierarchical
structure by combining p-values from lower levels using specified aggregation methods.}
\usage{
getGroupPvalues(pvals, groups, N, L, test)
}
\arguments{
\item{pvals}{NumericVector containing p-values for individual hypotheses}

\item{groups}{CharacterMatrix defining the group hierarchy structure}

\item{N}{Integer specifying the number of hypotheses}

\item{L}{Integer specifying the number of levels in the hierarchy}

\item{test}{CharacterVector specifying the aggregation method(s)}
}
\description{
Hierarchical Group P-value Calculation
The getGroupPvalues function computes aggregated p-values for each group in a hierarchical
structure by combining p-values from lower levels using specified aggregation methods.
}
