% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decomposition_step}
\alias{decomposition_step}
\title{Decomposition Step}
\usage{
decomposition_step(exp_mat_filename, data_dir)
}
\arguments{
\item{input}{filepath where merged expression matrix is stored, output directory to store data}
}
\value{
outputs one file with the shared component of expression per individual and C files for each specific expression component for each of the C contexts
}
\description{
Function to decompose expression into one shared component and specific components per context
}
