% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTL_mapping.R
\name{eQTL_mapping_step}
\alias{eQTL_mapping_step}
\title{eQTL Mapping Step}
\usage{
eQTL_mapping_step(
  SNP_file_name,
  snps_location_file_name,
  expression_file_name,
  gene_location_file_name,
  context,
  shared_specific,
  out_dir
)
}
\arguments{
\item{SNP_file_name}{\itemize{
\item full file path with genotypes of all individuals
}}

\item{snps_location_file_name}{\itemize{
\item full file path with snp ids, start, and end positions
}}

\item{expression_file_name}{\itemize{
\item full file path with expression matrix of individuals per gene
}}

\item{gene_location_file_name}{\itemize{
\item full file path with gene ids, start, and end positions
}}

\item{context}{\itemize{
\item name of context for output file naming purposes
}}

\item{shared_specific}{\itemize{
\item either "shared" if mapping eQTLs with shared component or "specific" if mapping eQTLs with specific component
}}

\item{out_dir}{\itemize{
\item full file path of output directory where eQTL file will be written out
}}
}
\value{
outputs one file of mapped cis eQTLs
}
\description{
Function to map cis eQTLs - cis window is defined as 1Mb
}
