% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulations}
\usage{
simulate_data(
  data_dir,
  N = 300,
  n_genes = 100,
  n_snps_per_gene = 1000,
  n_contexts = 10,
  maf = 0.2,
  w_corr = 0.2,
  missing = 0.05,
  hsq = rep(0.2, n_contexts),
  mus = rep(0, n_contexts),
  cisDist = 1e+06,
  seed = NULL
)
}
\arguments{
\item{data_dir}{\itemize{
\item full filepath of the directory where simulated data will be written to
}}

\item{N}{\itemize{
\item number of individuals simulated (must be a whole number)
}}

\item{n_genes}{\itemize{
\item number of genes simulated (must be a whole number)
}}

\item{n_snps_per_gene}{\itemize{
\item number of cis-SNPs per gene (must be a whole number)
}}

\item{n_contexts}{\itemize{
\item number of contexts to simulate (e.g. tissues or cell types) (must be a whole number)
}}

\item{maf}{\itemize{
\item minor allele frequency for genotypes
}}

\item{w_corr}{\itemize{
\item error covariance between contexts
}}

\item{missing}{\itemize{
\item decimal value signifying percentage of missingness in simulated expression data (e.g. parameter value of 0.3 would indicate 30\% missing values in outputted expression matrix)
}}

\item{hsq}{\itemize{
\item numeric vector with length of the number of contexts representing heritability explained by the eQTL in each context
}}

\item{mus}{\itemize{
\item numeric vector with length of the number of contexts representing average expression in each context.
}}

\item{cisDist}{\itemize{
\item cis window used to create geneloc and snploc files. Note: this value must be larger than 1
}}

\item{seed}{\itemize{
\item random seed for reproducibility. Default value is NULL
}}
}
\value{
outputs an expression matrix file, a genotype matrix file, a SNP location file, and a gene location file all in the format needed for FastGxC's decomposition step and then subsequent eQTL mapping step with Matrix eQTL.
}
\description{
Function to generate simulated SNP and expression data
}
\examples{
\dontrun{
simulate_data(
  data_dir = tempdir(),
  N = 10,
  n_genes = 5,
  n_snps_per_gene = 10,
  n_contexts = 3,
  seed = 42
)
}

}
