% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulations}
\usage{
simulate_data(
  data_dir,
  N = 300,
  n_genes = 100,
  n_snps_per_gene = 100,
  n_contexts = 5,
  maf = 0.1,
  w_corr = 0.2,
  v_e = 1,
  sim_scenario = "single-context_het"
)
}
\arguments{
\item{data_dir}{\itemize{
\item directory where simulated data is written out to
}}

\item{N}{\itemize{
\item number of individuals simulated
}}

\item{n_genes}{\itemize{
\item number of genes simulated
}}

\item{n_snps_per_gene}{\itemize{
\item number of cis-SNPs per gene
}}

\item{n_contexts}{\itemize{
\item number of contexts to simulate (e.g. tissues or cell types)
}}

\item{maf}{\itemize{
\item minor allele frequency for genotypes
}}

\item{w_corr}{\itemize{
\item error covariance between contexts
}}

\item{v_e}{\itemize{
\item error variance in each context (maybe take this out and set it to 1)
}}

\item{sim_scenario}{\itemize{
\item either "null", "single-context_het", or "multi-context_het" to signify simulations under the null case (no genetic effects in any context), the case of single context heterogeneity (one context drives the genetic effect heterogeneity), or the case of multi-context heterogeneity (heterogeneity of genetic effects across multiple contexts)
}}
}
\value{
outputs an expression matrix file, a genotype matrix file, a SNP location file, and a gene location file all in the format needed for FastGxC's decomposition step and then subsequent eQTL mapping step with Matrix eQTL.
}
\description{
Function to generate simulated SNP and expression data
}
