% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulations}
\usage{
simulate_data(
  data_dir,
  N = 300,
  n_genes = 100,
  n_snps_per_gene = 10000,
  n_contexts = 50,
  maf = 0.1,
  w_corr = 0.2,
  v_e = 1,
  missing = 0,
  seed = 1,
  sim_scenario = "single_context_het"
)
}
\arguments{
\item{data_dir}{\itemize{
\item full filepath of the directory where simulated data will be written to
}}

\item{N}{\itemize{
\item number of individuals simulated (must be a whole number)
}}

\item{n_genes}{\itemize{
\item number of genes simulated (must be a whole number)
}}

\item{n_snps_per_gene}{\itemize{
\item number of cis-SNPs per gene (must be a whole number)
}}

\item{n_contexts}{\itemize{
\item number of contexts to simulate (e.g. tissues or cell types) (must be a whole number)
}}

\item{maf}{\itemize{
\item minor allele frequency for genotypes
}}

\item{w_corr}{\itemize{
\item error covariance between contexts
}}

\item{v_e}{\itemize{
\item error variance in each context (maybe take this out and set it to 1)
}}

\item{missing}{\itemize{
\item decimal value signifying percentage of missingness in simulated expression data (e.g. parameter value of 0.3 would indicate 30\% missing values in outputted expression matrix)
}}

\item{seed}{\itemize{
\item can set seed for reproducibility
}}

\item{sim_scenario}{\itemize{
\item must be either "null" or "single_context_het" to signify simulations under the null case (no genetic effects in any context) or the case of single context heterogeneity (one context drives the genetic effect heterogeneity)
}}
}
\value{
outputs an expression matrix file, a genotype matrix file, a SNP location file, and a gene location file all in the format needed for FastGxC's decomposition step and then subsequent eQTL mapping step with Matrix eQTL.
}
\description{
Function to generate simulated SNP and expression data
}
