% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeBH.R
\name{to_TreeBH_input}
\alias{to_TreeBH_input}
\title{Converting eQTL Mapping Output to TreeBH Input}
\usage{
to_TreeBH_input(data_dir, shared_file, context_names, out_dir)
}
\arguments{
\item{data_dir}{\itemize{
\item full filepath of the directory where eQTL output files are stored. This function assumes that files are named as outputted by FastGxC's eQTL mapping function
}}

\item{shared_file}{\itemize{
\item full filepath to the shared cis_pairs file from the output of the mapping function
}}

\item{context_names}{\itemize{
\item vector of all context names in the format c("tissue1", "tissue2", ..., etc.)
}}

\item{out_dir}{\itemize{
\item full filepath of the output directory where the input of TreeBH can be stored
}}
}
\value{
A data.frame with columns:
\itemize{
\item gene_name     : gene name
\item snp_name      : snp name
\item context_name  : context name (or "shared")
\item gene_snp      : concatenated "gene_SNP"
\item component_tag : "gene_SNP_specific" or "gene_SNP_shared"
\item context_tag   : "gene_SNP_component_context"
\item p_value       : numeric p-value
}
}
\description{
Converting eQTL Mapping Output to TreeBH Input
}
\examples{
\dontrun{
to_TreeBH_input(
  data_dir = tempdir(),
  shared_file = file.path(tempdir(), "shared_specific.cis_pairs.txt"),
  context_names = c("context1", "context2", "context3"),
  out_dir = tempdir()
)
}

}
