% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeBH.R
\name{treeBH_step}
\alias{treeBH_step}
\title{Multiple Testing Correction with TreeBH
Function to adjust for hierarchical multiple testing correction using TreeBH.}
\usage{
treeBH_step(matrix, fdr_thres, out_dir)
}
\arguments{
\item{matrix}{\itemize{
\item output of to_TreeBH_Input, matrix with the groups and pvalues
}}

\item{fdr_thres}{\itemize{
\item value between 0 and 1 that signifies what FDR threshold for multiple testing correction. The same value will be used across all hierarchical levels.
}}

\item{out_dir}{\itemize{
\item full filepath of the output directory where the output of TreeBH can be stored
}}
}
\value{
A matrix with identification columns and output columns:
\itemize{
\item gene                    : gene name
\item snp                     : snp name
\item context                 : context name or "specific"
\item treeBH Output Columns   : eGene, eQTL, Component, Context Specific
}
}
\description{
Multiple Testing Correction with TreeBH
Function to adjust for hierarchical multiple testing correction using TreeBH.
}
