% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeQTL.R
\name{treeQTL_step}
\alias{treeQTL_step}
\title{Multiple Testing Correction}
\usage{
treeQTL_step(
  data_dir,
  snps_location_file_name,
  gene_location_file_name,
  out_dir,
  fdr_thresh = 0.05
)
}
\arguments{
\item{data_dir}{\itemize{
\item full filepath of the directory where eQTL output files are stored. This function assumes that files are named as outputted by FastGxC's eQTL mapping function
}}

\item{snps_location_file_name}{\itemize{
\item full filepath of the snpsloc file used in the eQTL mapping step
}}

\item{gene_location_file_name}{\itemize{
\item full filepath of the geneloc file used in the eQTL mapping step
}}

\item{out_dir}{\itemize{
\item full filepath of the output directory that FDR adjusted eQTLs should be written out to.
}}

\item{fdr_thresh}{\itemize{
\item value between 0 and 1 that signifies what FDR threshold for multiple testing correction. The same value will be used across all hierarchical levels.
}}
}
\value{
outputs one file of specific eGenes across all contexts and one file of shared eGenes. Outputs an eAssociation file for each context and one for shared eQTLs with snp-gene pairs and FDR adjusted p-values.
}
\description{
Function to adjust for hierarchical multiple testing correction using TreeQTL. Runs multiple testing correction for both FastGxC shared and specific eQTLs.
}
